TurnArrowTests -> ManeuverViewTests
