+++ MapboxNavigationTests/ReferenceImages_64/MapboxNavigationTests.ManeuverViewTests/testStraightRoundabout@3x.png
